# EnigmaStepThree/openapi.yaml
openapi: 3.0.3
info:
  title: Get One Enigma Step API
  description: API to get one step to resolve the enigma of how to put a giraffe into a refrigerator
  contact:
    email: afortega@bancolombia.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
  
servers:
  
  - description: Local server
    url: 'localhost:{port}/v2'
    variables:
      port:
        enum:
          - '443'
          - '8080'
        default: '8080'
      protocol:
        enum:
          - 'http'
          - 'https'
        default: 'https'
        
tags:
  - name: user
    description: Operations about user
paths:
  /getStep:
    post:
      tags:
        - "get-one-enigma-step-API"
      summary: Get one engima step API
      description: ""
      operationId: getStep
      requestBody:
        description: Get one enigma step API
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonApiBodyRequest'
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JsonApiBodyResponseSuccess'
        '424':
          description: bad input parameter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JsonApiBodyResponseErrors'


components:
  schemas:
    JsonApiBodyRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/GetEnigmaRequest'
    GetEnigmaRequest:
      type: object
      required:
        - header
        - enigma
      properties:
        header:
          $ref: '#/components/schemas/Header'
        enigma:
          type: string
    Header:
      type: object
      required:
        - id
        - type
      properties:
        id:
          type: string
        type:
          type: string      
    JsonApiBodyResponseErrors:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - code
        - detail
        - id
        - source
        - status
        - title
      properties:
        code:
          type: string
        detail:
          type: string
        id:
          type: string
        source:
          type: string
        status:
          type: string
        title:
          type: string
    JsonApiBodyResponseSuccess:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/GetEnigmaStepResponse'
    GetEnigmaStepResponse:
      type: object
      required:
        - header
        - answer
      properties:
        header:
          $ref: '#/components/schemas/Header'
        answer:
          type: string